Refactor: DTO들을 모델타입 외부로 분리
