+++ ProjectManagerServer/Sources/App/Models/DeleteProjectItem.swift
@@ -1,0 +1,12 @@
+//
+//  DeleteProjectItem.swift
+//  
+//
+//  Created by Ryan-Son on 2021/07/08.
+//
+
+import Vapor
+
+struct DeleteProjectItem: Content {
+    let id: UUID
+}
